Fix DB 3

+++ database/migrations/kNet_DB/2026_02_18_091753_alter_table_budgana.php
@@ -14,12 +14,12 @@ class AlterTableBudgana extends Migration
     public function up()
     {
         Schema::table('u_budgana', function (Blueprint $table) {
-            $table->float('u_budg1',15,2)->nullable()->default(0.0)->change();
-            $table->float('u_budg2', 15, 2)->nullable()->default(0.0)->change();
-            $table->float('u_budg3', 15, 2)->nullable()->default(0.0)->change();
-            $table->float('u_kobudg1', 15, 2)->nullable()->default(0.0)->change();
-            $table->float('u_kobudg2', 15, 2)->nullable()->default(0.0)->change();
-            $table->float('u_kobudg3', 15, 2)->nullable()->default(0.0)->change();
+            $table->decimal('u_budg1',15,2)->nullable()->default(0.0)->change();
+            $table->decimal('u_budg2', 15, 2)->nullable()->default(0.0)->change();
+            $table->decimal('u_budg3', 15, 2)->nullable()->default(0.0)->change();
+            $table->decimal('u_kobudg1', 15, 2)->nullable()->default(0.0)->change();
+            $table->decimal('u_kobudg2', 15, 2)->nullable()->default(0.0)->change();
+            $table->decimal('u_kobudg3', 15, 2)->nullable()->default(0.0)->change();
         });
     }
